# Improve performance by using bytesize when comparing if the string changed